+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/data/quarter/source/SettingsDataSource.kt
@@ -2,5 +2,5 @@ package com.gdavidpb.tuindice.record.data.quarter.source
 
 interface SettingsDataSource {
 	suspend fun isGetQuartersOnCooldown(): Boolean
-	suspend fun setGetQuartersCooldown()
+	suspend fun setGetQuartersOnCooldown()
 }
