+++ summary/src/main/kotlin/com/gdavidpb/tuindice/summary/data/account/source/LocalDataSource.kt
@@ -4,9 +4,6 @@ import com.gdavidpb.tuindice.base.domain.model.Account
 import kotlinx.coroutines.flow.Flow
 
 interface LocalDataSource {
-	suspend fun isGetAccountOnCooldown(): Boolean
-	suspend fun setGetAccountCooldown()
-
 	suspend fun getAccount(uid: String): Flow<Account?>
 	suspend fun saveAccount(uid: String, account: Account)
