+++ summary/src/main/kotlin/com/gdavidpb/tuindice/summary/data/account/source/SettingsDataSource.kt
@@ -1,0 +1,6 @@
+package com.gdavidpb.tuindice.summary.data.account.source
+
+interface SettingsDataSource {
+	suspend fun isGetAccountOnCooldown(): Boolean
+	suspend fun setGetAccountOnCooldown()
+}
